refactor: refactor file upload modal
